+++ magento/client.py
@@ -234,7 +234,7 @@ class Magento(APISession):
     # Carts
     # =====
 
-    def get_cart(self, *, cart_id: PathId, **kwargs) -> MagentoEntity:
+    def get_cart(self, cart_id: PathId, **kwargs) -> MagentoEntity:
         return self.get_json_api(f"/V1/carts/{cart_id}", **kwargs)
 
     def get_carts(self, *, query: Query = None, limit=-1, **kwargs) -> Iterable[MagentoEntity]: