+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `get_order_item`
 * `get_cart` now accepts the cart ID as a positional argument
+* Escape path arguments in all calls
 
 ## 1.11.0 (2024/04/12)