+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Add `get_order_item`, `get_customer("me")`
+* Add `get_current_customer`, `activate_current_customer`, `change_current_customer_password`
 * `get_cart` now accepts the cart ID as a positional argument
 * Expose `magento.types.*` under `magento`
 * Escape path arguments in all calls